requirements.txt
python-dotenv
requests
urllib3
flask
functions-framework